refactor: separate failed dialog in its own file
