+++ lib/pages/farm/components/my_farm_item.dart
@@ -6,6 +6,7 @@ import 'package:ax_dapp/pages/farm/dialogs/unstake_dialog.dart';
 import 'package:ax_dapp/pages/farm/modules/box_decoration.dart';
 import 'package:ax_dapp/pages/farm/modules/page_text_style.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
+import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -183,19 +184,25 @@ Widget myFarmItem(
                               ? farm.strStakedAlias.value
                               : farm.strStakedSymbol.value,
                         ),
-                    await farm.claim(),
-                    showDialog<void>(
-                      context: context,
-                      builder: rewardClaimDialog,
-                    ),
-                    context.read<TrackingCubit>().onClaimRewardsSuccess(
-                          tickerPair: farm.athlete == null
-                              ? farm.strName
-                              : farm.athlete!,
-                          tickerPairName: farm.strStakedAlias.value.isNotEmpty
-                              ? farm.strStakedAlias.value
-                              : farm.strStakedSymbol.value,
-                        ),
+                    await farm.claim().then((value) {
+                      showDialog<void>(
+                        context: context,
+                        builder: rewardClaimDialog,
+                      );
+                      context.read<TrackingCubit>().onClaimRewardsSuccess(
+                            tickerPair: farm.athlete == null
+                                ? farm.strName
+                                : farm.athlete!,
+                            tickerPairName: farm.strStakedAlias.value.isNotEmpty
+                                ? farm.strStakedAlias.value
+                                : farm.strStakedSymbol.value,
+                          );
+                    }).catchError((error) {
+                      showDialog<void>(
+                        context: context,
+                        builder: (context) => const FailedDialog(),
+                      );
+                    })
                   },
                   child: Text(
                     'Claim Rewards',
