+++ lib/pages/trade/components/trade_approve_button.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: file_names
 
 import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
