+++ lib/service/approve_button.dart
@@ -1,4 +1,5 @@
 import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:flutter/material.dart';
 
 // This code changes the state of the button
